ch4/ucx: move multivci code to ucx_vci.c
All multivci related initialization code in ucx_vci.c.
